Rename missleading tag class
