+++ src/main/java/tech/jhipster/lite/module/domain/resource/InvalidJHipsterModuleTagException.java
@@ -1,4 +1,4 @@
-package tech.jhipster.lite.module.domain;
+package tech.jhipster.lite.module.domain.resource;
 
 import tech.jhipster.lite.error.domain.GeneratorException;
