+++ src/main/java/tech/jhipster/lite/module/domain/resource/JHipsterModuleApiDoc.java
@@ -2,13 +2,13 @@ package tech.jhipster.lite.module.domain.resource;
 
 import tech.jhipster.lite.error.domain.Assert;
 
-public record JHipsterModuleApiDoc(JHipsterModuleTag tag, JHipsterModuleOperation operation) {
-  public JHipsterModuleApiDoc(String tag, String operation) {
-    this(new JHipsterModuleTag(tag), new JHipsterModuleOperation(operation));
+public record JHipsterModuleApiDoc(JHipsterModuleGroup group, JHipsterModuleOperation operation) {
+  public JHipsterModuleApiDoc(String group, String operation) {
+    this(new JHipsterModuleGroup(group), new JHipsterModuleOperation(operation));
   }
 
   public JHipsterModuleApiDoc {
-    Assert.notNull("tag", tag);
+    Assert.notNull("group", group);
     Assert.notNull("operation", operation);
   }
 }
