+++ src/main/java/tech/jhipster/lite/module/domain/resource/JHipsterModuleGroup.java
@@ -1,0 +1,18 @@
+package tech.jhipster.lite.module.domain.resource;
+
+import java.util.List;
+import tech.jhipster.lite.error.domain.Assert;
+
+public record JHipsterModuleGroup(String group) {
+  public JHipsterModuleGroup {
+    Assert.field("group", group).notBlank().maxLength(50);
+  }
+
+  public String get() {
+    return group();
+  }
+
+  public List<String> list() {
+    return List.of(group());
+  }
+}
