+++ src/main/java/tech/jhipster/lite/module/domain/resource/JHipsterModuleResource.java
@@ -3,7 +3,6 @@ package tech.jhipster.lite.module.domain.resource;
 import tech.jhipster.lite.error.domain.Assert;
 import tech.jhipster.lite.module.domain.JHipsterModuleFactory;
 import tech.jhipster.lite.module.domain.JHipsterModuleSlug;
-import tech.jhipster.lite.module.domain.JHipsterModuleTags;
 
 public class JHipsterModuleResource {
