+++ src/main/java/tech/jhipster/lite/module/infrastructure/primary/RestJHipsterModule.java
@@ -4,8 +4,8 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.ArrayList;
 import java.util.Collection;
 import tech.jhipster.lite.error.domain.Assert;
-import tech.jhipster.lite.module.domain.JHipsterModuleTag;
 import tech.jhipster.lite.module.domain.resource.JHipsterModuleResource;
+import tech.jhipster.lite.module.domain.resource.JHipsterModuleTag;
 
 @Schema(name = "JHipsterModule", description = "Information for a JHipster module")
 class RestJHipsterModule {
