+++ src/test/java/tech/jhipster/lite/module/domain/JHipsterTagsTest.java
@@ -5,7 +5,8 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.Test;
 import tech.jhipster.lite.UnitTest;
 import tech.jhipster.lite.error.domain.StringWithWitespacesException;
-import tech.jhipster.lite.module.domain.JHipsterModuleTags.JHipsterModuleTagsBuilder;
+import tech.jhipster.lite.module.domain.resource.JHipsterModuleTags;
+import tech.jhipster.lite.module.domain.resource.JHipsterModuleTags.JHipsterModuleTagsBuilder;
 
 @UnitTest
 class JHipsterTagsTest {
